feat: implement authorisation

+++ app/Http/Controllers/IdeaController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StoreIdeaRequest;
 use App\Http\Requests\UpdateIdeaRequest;
 use App\IdeaStatus;
 use App\Models\Idea;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -82,6 +83,9 @@ class IdeaController extends Controller
      */
     public function show(Idea $idea)
     {
+        // METHOD 1: in Controller
+        Gate::authorize('workWith', $idea);
+
         return view('idea.show', ['idea' => $idea]);
     }
 
@@ -98,7 +102,7 @@ class IdeaController extends Controller
      */
     public function update(UpdateIdeaRequest $request, Idea $idea)
     {
-        //
+        Gate::authorize('workWith', $idea);
     }
 
     /**
@@ -107,6 +111,7 @@ class IdeaController extends Controller
     public function destroy(Idea $idea)
     {
         // authorise that this is allowed
+        Gate::authorize('workWith', $idea);
         $idea->delete();
 
         // return redirect()->route('idea.index');
